[HW12] Work version maven

+++ service/src/main/java/com/senla/carservice/OrderServiceImpl.java
@@ -3,17 +3,17 @@ package com.senla.carservice;
 import com.senla.carservice.annotation.Singleton;
 import com.senla.carservice.objectadjuster.dependencyinjection.annotation.Dependency;
 import com.senla.carservice.objectadjuster.propertyinjection.annotation.ConfigProperty;
-import com.senla.carservice.dao.MasterDao;
-import com.senla.carservice.dao.OrderDao;
-import com.senla.carservice.dao.PlaceDao;
-import com.senla.carservice.dao.connection.DatabaseConnection;
-import com.senla.carservice.domain.Master;
-import com.senla.carservice.domain.Order;
-import com.senla.carservice.domain.Place;
-import com.senla.carservice.domain.enumaration.Status;
+import com.senla.carservice.MasterDao;
+import com.senla.carservice.OrderDao;
+import com.senla.carservice.PlaceDao;
+import com.senla.carservice.connection.DatabaseConnection;
+import com.senla.carservice.Master;
+import com.senla.carservice.Order;
+import com.senla.carservice.Place;
+import com.senla.carservice.enumaration.Status;
 import com.senla.carservice.exception.BusinessException;
-import com.senla.carservice.service.enumaration.SortParameter;
-import com.senla.carservice.util.DateUtil;
+import com.senla.carservice.enumaration.SortParameter;
+import com.senla.carservice.DateUtil;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -30,6 +30,8 @@ public class OrderServiceImpl implements OrderService {
     private MasterDao masterDao;
     @ConfigProperty
     private boolean isBlockShiftTime;
+    @ConfigProperty
+    private boolean isBlockDeleteOrder;
     @Dependency
     private DatabaseConnection databaseConnection;
 
@@ -217,6 +219,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void deleteOrder(Order order) {
+        if (isBlockDeleteOrder) {
+            throw new BusinessException("Permission denied");
+        }
         try {
             databaseConnection.disableAutoCommit();
             orderDao.deleteRecord(order, databaseConnection);
@@ -231,11 +236,11 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void shiftLeadTime(Order order, Date executionStartTime, Date leadTime) {
+        if (isBlockShiftTime) {
+            throw new BusinessException("Permission denied");
+        }
         try {
             databaseConnection.disableAutoCommit();
-            if (isBlockShiftTime) {
-                throw new BusinessException("Permission denied");
-            }
             DateUtil.checkDateTime(executionStartTime, leadTime, false);
             checkStatusOrderShiftTime(order);
             order.setLeadTime(leadTime);
